+++ src/data_conversion.rs
@@ -7,7 +7,7 @@ use crate::{
     app::{data_farmer::DataCollection, AxisScaling},
     canvas::tui_widgets::time_chart::Point,
     data_collection::{cpu::CpuDataType, memory::MemHarvest, temperature::TemperatureType},
-    utils::{data_prefixes::*, data_units::DataUnit, gen_util::*},
+    utils::{data_prefixes::*, data_units::DataUnit, general::*},
     widgets::{DiskWidgetData, TempWidgetData},
 };
